Provide transposed registry to exporter
Revert earlier ExportData workarounds for server
